AGL-3215: Classificatie als legende
aangepast: laagstijleditor component opgesplitst in verschillende,
kleinere bestanden.
